feat: remove ormconfig.json
