--- ormconfig.json
@@ -1,18 +1,0 @@
-{
-  "type": "postgres",
-  "host": "localhost",
-  "port": "5432",
-  "username": "postgres",
-  "password": "qwerty3355",
-  "database": "fastify_db",
-  "synchronize": false,
-  "logging": false,
-  "entities": ["src/resources/**/*{.js,.ts}"],
-  "migrations": ["src/migration/**/*{.js,.ts}"],
-  "subscribers": ["src/subscriber/**/*{.js,.ts}"],
-  "cli": {
-    "entitiesDir": "src/resources",
-    "migrationsDir": "src/migration",
-    "subscribersDir": "src/subscriber"
-  }
-}
