+++ package.json
@@ -16,8 +16,9 @@
     "start:prod": "nodemon -L --exitcrash dist/server.js",
     "test": "cross-env DEBUG=rs:* jest --testMatch \"<rootDir>/test/e2e/test/*.test.js\" --noStackTrace --runInBand",
     "test:auth": "cross-env DEBUG=rs:* TEST_MODE=auth jest --noStackTrace",
-    "typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js",
-    "typeorm:run": "node --require ts-node/register ./node_modules/typeorm/cli.js migration:run"
+    "typeorm": "node --require ./node_modules/ts-node/register ./node_modules/.bin/typeorm",
+    "typeorm:generate": "npm run typeorm -- migration:generate --config src/serverOptions/ormmigration.ts --name",
+    "typeorm:run": "npm run typeorm -- migration:run --config src/serverOptions/ormmigration.ts"
   },
   "keywords": [
     "rs",
